requirements.txt
supabase>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.25.0,<1.0.0
pydantic>=2.5.0,<3.0.0
loguru>=0.7.0,<1.0.0
postgrest>=0.13.0,<1.0.0